Changes and fixes in shell, ps, processHandler, strcpy, strncpy
Fixed shell \n. Added total processes to ps and updated it to 200, added return value on killProcess, added return value on an unsuccessful newProcess, fixed strcpy and strncpy in kernel, processes' array is now dynamic

+++ TPE/Userland/SampleCodeModule/stdlib.c
@@ -101,7 +101,7 @@ int itoa(int value, char* buffer, int base){
 /*
  * Convierte  string a int 
  */
-int atoi(char * s){ 		
+int64_t atoi(char * s){ 		
 	int i = 0;
 	int num = 0;
 	while(s[i]){
